Move datarobot_otel_metrics into core/telemetry/ package
Aligns the metrics bootstrap with main's new core/telemetry/ package
(agent/datarobot_otel/memory/nat_*). The module is stdlib-only at import
time (opentelemetry imports are deferred inside bootstrap_metrics_provider)
and core/telemetry/__init__.py is empty, so this does not pull agent
dependencies into importers.
